+++ packages/page-sbt/src/Top.tsx
@@ -3,41 +3,40 @@
 
 // import type { VerifiableCredential } from '@zcloak/vc/types';
 
-import { Box, Container, Typography } from '@mui/material';
+import { Box, Container, Stack, Typography } from '@mui/material';
 import React from 'react';
 
-// import { CTypeName } from '@credential/react-components';
-// import { DidName } from '@credential/react-dids';
+import { ConnectWallet, EthWalletAddress, Network, useAccount } from '@credential/react-components';
 
-// interface Props {
-//   vc: VerifiableCredential<boolean>;
-// }
-
-// function Item({ label, value }: { label: string; value?: React.ReactNode }) {
-//   return (
-//     <Stack
-//       bgcolor='#FBFAFC'
-//       border='1px solid #EBEAED'
-//       direction='row'
-//       display='inline-flex'
-//       lineHeight={'18px'}
-//       maxWidth={520}
-//       paddingX={2}
-//       paddingY={1}
-//     >
-//       <Box sx={{ width: 90 }}>{label}</Box>
-//       <Box sx={{ color: '#5D5D5B' }}>{value}</Box>
-//     </Stack>
-//   );
-// }
+import Faucet from './Faucet';
 
 function Top() {
+  const { isConnected } = useAccount();
+
   return (
     <Box sx={{ bgcolor: 'common.white', paddingTop: '36px', paddingBottom: '20px' }}>
-      <Container maxWidth='xl' sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
+      <Container
+        maxWidth='xl'
+        sx={{ display: 'flex', flexDirection: 'row', gap: 2, justifyContent: 'space-between', alignItems: 'center' }}
+      >
         <Typography variant='h1'>Mint zkID Card</Typography>
-        {/* <Item label='cType:' value={<CTypeName cTypeHash={vc.ctype} />} /> */}
-        {/* <Item label='Attester:' value={<DidName shorten={false} value={vc.issuer} />} /> */}
+        {isConnected ? (
+          <Stack
+            direction={{
+              sm: 'row',
+              xs: 'column'
+            }}
+            spacing={2}
+          >
+            <Faucet />
+            <Network />
+            <EthWalletAddress />
+          </Stack>
+        ) : (
+          <ConnectWallet sx={{ width: 200 }} variant='outlined'>
+            Connect Wallet
+          </ConnectWallet>
+        )}
       </Container>
     </Box>
   );